added countlambdas, and mintime etc to lambdas map

+++ scripts_cpp/causets_cpp/embeddedcauset.h
@@ -132,11 +132,11 @@ class EmbeddedCauset: public Causet
 
 
         std::map<int,std::vector<int>> get_lambdas(double& t_f, double r_S=2);
-        std::map<int,int>            count_lambdas(double& t_f, double r_S=2);
+        std::map<int,double>         count_lambdas(double& t_f, double r_S=2);
 
 
         std::map<int,std::vector<int>> get_HRVs(double& t_f, double r_S = 2);
-        std::map<int,int>            count_HRVs(double& t_f, double r_S = 2);
+        std::map<int,double>         count_HRVs(double& t_f, double r_S = 2);
 
 
         void save_molecules(const char* path_file_ext = "boh", 
@@ -149,13 +149,13 @@ class EmbeddedCauset: public Causet
 
         std::map<int,std::vector<int>> get_lambdas_from_futlinks(double& t_f,
                                                                  double r_S=2);
-        std::map<int,int> get_lambdas_sizes(double& t_f, double r_S = 2);
-        std::map<int,int> get_lambdas_distr(const std::map<int, int> & lambdas);
+        std::map<int,double> get_lambdas_sizes(double& t_f, double r_S = 2);
+        std::map<int,double> get_lambdas_distr(const std::map<int, double> & lambdas);
 
         std::map<int,std::vector<int>> get_HRVs_from_futlinks(double& t_f,
                                                               double r_S=2);
-        std::map<int,int> get_HRVs_distr_from_futlinks(double& t_f, 
-                                                       double r_S = 2);
+        std::map<int,double> get_HRVs_distr_from_futlinks(double& t_f, 
+                                                          double r_S = 2);
 
         //Destructor
         ~EmbeddedCauset();
